provisioning/orchestrator/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0

# Docker
docker==7.0.0

# XenServer (utiliser XenAPI ou requests pour l'API REST)
requests==2.31.0

# Cloud-init generation
pyyaml==6.0.1
jinja2==3.1.3

# Logging
python-json-logger==2.0.7

# Utils
python-dateutil==2.8.2
